requirements.txt
galaxy.plugin.api==0.66.0
psutil==5.7.0